made the grid support int32

+++ Obstacles/Fence.cs
@@ -1,8 +1,8 @@
 namespace Obstacles;
 class Fence : Obstacle {
-    private byte _endX;
-    private byte _endY;
-    public Fence(byte startX, byte startY, byte endX, byte endY) : base(startX, startY) {
+    private int _endX;
+    private int _endY;
+    public Fence(int startX, int startY, int endX, int endY) : base(startX, startY) {
         _endX = endX;
         _endY = endY;
     }
